moved SimuComConfig to de.uka.ipd.sdq.simucomframework.core
