+++ bundles/de.uka.ipd.sdq.simucomframework/META-INF/MANIFEST.MF
@@ -40,6 +40,7 @@ Require-Bundle: de.uka.ipd.sdq.simucomframework.variables,
  org.palladiosimulator.measurementframework;bundle-version="1.0.0",
  org.palladiosimulator.pcm.edp2.measuringpoint;bundle-version="1.0.0",
  org.palladiosimulator.edp2.util;bundle-version="1.0.0",
- org.palladiosimulator.pcm;bundle-version="4.1.0"
+ org.palladiosimulator.pcm;bundle-version="4.1.0",
+ de.uka.ipd.sdq.simucomframework.core;bundle-version="5.2.1";visibility:=reexport
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
