+++ bundles/org.palladiosimulator.simucom.reliability/src/org/palladiosimulator/simucom/reliability/jobs/ReliabilityExtensionTab.java
@@ -13,7 +13,7 @@ import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Group;
 
-import de.uka.ipd.sdq.simucomframework.SimuComConfig;
+import de.uka.ipd.sdq.simucomframework.core.SimuComConfig;
 
 /**
  * This class defines a tab with reliability-specific configuration options.
