+++ tests/de.uka.ipd.sdq.codegen.simucontroller.tests/src/de/uka/ipd/sdq/codegen/simucontroller/tests/SimuControllerTestSuite.java
@@ -21,7 +21,7 @@ import org.palladiosimulator.analyzer.workflow.core.ConstantsContainer;
 import org.palladiosimulator.recorderframework.sensorframework.SensorFrameworkRecorderConfiguration;
 import org.palladiosimulator.recorderframework.sensorframework.SensorFrameworkRecorderConfigurationFactory;
 
-import de.uka.ipd.sdq.simucomframework.SimuComConfig;
+import de.uka.ipd.sdq.simucomframework.core.SimuComConfig;
 
 /**
  *
