+++ tests/de.uka.ipd.sdq.simucomframework.variables.tests/src/de/uka/ipd/sdq/simucomframework/variables/tests/StoExVisitorTests.java
@@ -12,7 +12,7 @@ import org.apache.log4j.PatternLayout;
 
 import de.uka.ipd.sdq.probfunction.math.IProbabilityFunctionFactory;
 import de.uka.ipd.sdq.probfunction.math.impl.ProbabilityFunctionFactoryImpl;
-import de.uka.ipd.sdq.simucomframework.SimuComDefaultRandomNumberGenerator;
+import de.uka.ipd.sdq.simucomframework.core.SimuComDefaultRandomNumberGenerator;
 import de.uka.ipd.sdq.simucomframework.variables.StackContext;
 import de.uka.ipd.sdq.simucomframework.variables.cache.StoExCache;
 import de.uka.ipd.sdq.simucomframework.variables.functions.MaxDeviationFunction;
